+++ rat-text/src/text_area.rs
@@ -36,6 +36,8 @@ use std::ops::Range;
 
 /// Text area widget.
 ///
+/// [Example](https://github.com/thscharler/rat-salsa/blob/master/rat-text/examples/textarea2.rs)
+///
 /// Backend used is [ropey](https://docs.rs/ropey/latest/ropey/), so large
 /// texts are no problem. Editing time increases with the number of
 /// styles applied. Everything below a million styles should be fine.
@@ -54,9 +56,8 @@ use std::ops::Range;
 /// Scrolling doesn't depend on the cursor, but the editing and move
 /// functions take care that the cursor stays visible.
 ///
-/// Wordwrap is not available. For display only use
-/// [Paragraph](https://docs.rs/ratatui/latest/ratatui/widgets/struct.Paragraph.html), as
-/// for editing: why?
+/// Text wrapping is available, hard line-breaks at the right margin,
+/// or decent word-wrapping.
 ///
 /// You can directly access the underlying Rope for readonly purposes, and
 /// conversion from/to byte/char positions are available. That should probably be
@@ -182,10 +183,17 @@ impl Clone for TextAreaState {
 pub enum TextWrap {
     /// Don't break, shift text to the left.
     Shift,
-    /// Hard break at the width.
+    /// Hard break at the right border.
     Hard,
-    /// Word break within the given right margin.
-    /// If not possible do a hard break.
+    /// Wraps the text at word boundaries.
+    ///
+    /// The parameter gives an area before the right border where
+    /// breaks are preferred. The first space that falls in this
+    /// region will break. Otherwise, the last space before will be
+    /// used, or the word will be hard-wrapped.
+    ///
+    /// Space is the word-separator. Words will be broken if they
+    /// contain a hyphen, a soft-hyphen or a zero-width-space.
     Word(u16),
 }
 
@@ -617,34 +625,47 @@ impl TextAreaState {
         self.value.expand_tabs()
     }
 
-    /// Show control characters.
+    /// Show glyphs for control characters.
     #[inline]
     pub fn set_show_ctrl(&mut self, show_ctrl: bool) {
         self.value.set_glyph_ctrl(show_ctrl);
     }
 
-    /// Show control characters.
+    /// Show glyphs for control characters.
     pub fn show_ctrl(&self) -> bool {
         self.value.glyph_ctrl()
     }
 
-    /// Show glyphs for text-wrap.
+    /// Show glyphs for text-wrapping.
+    /// Shows inserted line-breaks, zero-width space (U+200B) or the soft hyphen (U+00AD).
     #[inline]
     pub fn set_wrap_ctrl(&mut self, wrap_ctrl: bool) {
         self.value.set_wrap_ctrl(wrap_ctrl);
     }
 
-    /// Show control characters.
+    /// Show glyphs for text-wrapping.
+    /// Shows inserted line-breaks, zero-width space (U+200B) or the soft hyphen (U+00AD).
     pub fn wrap_ctrl(&self) -> bool {
         self.value.wrap_ctrl()
     }
 
-    /// Text breaking.
+    /// Text wrapping mode.
+    ///
+    /// * TextWrap::Shift means no wrapping.
+    /// * TextWrap::Hard hard-wraps at the right border.
+    /// * TextWrap::Word(n) does word wrapping.
+    ///   n gives the size of the break-region close to the right border.
+    ///   The first space that falls in this region is taken as a break.
+    ///   If that is not possible this will break at the last space before.
+    ///   If there is no space it will hard-wrap the word.
+    ///
+    ///   Space is used as word separator. Hyphen will be used to break
+    ///   a word, and soft-hyphen and zero-width-space will be recognized too.
     pub fn set_text_wrap(&mut self, text_wrap: TextWrap) {
         self.text_wrap = text_wrap;
     }
 
-    /// Text breaking.
+    /// Text wrapping.
     pub fn text_wrap(&self) -> TextWrap {
         self.text_wrap
     }
@@ -688,7 +709,7 @@ impl TextAreaState {
         self.value.clipboard()
     }
 
-    /// Copy to internal buffer
+    /// Copy selection to clipboard.
     #[inline]
     pub fn copy_to_clip(&mut self) -> bool {
         let Some(clip) = self.value.clipboard() else {
@@ -699,7 +720,7 @@ impl TextAreaState {
         false
     }
 
-    /// Cut to internal buffer
+    /// Cut selection to clipboard.
     #[inline]
     pub fn cut_to_clip(&mut self) -> bool {
         let Some(clip) = self.value.clipboard() else {
@@ -712,7 +733,7 @@ impl TextAreaState {
         }
     }
 
-    /// Paste from internal buffer.
+    /// Paste from clipboard.
     #[inline]
     pub fn paste_from_clip(&mut self) -> bool {
         let Some(clip) = self.value.clipboard() else {
@@ -728,7 +749,12 @@ impl TextAreaState {
 }
 
 impl TextAreaState {
-    /// Set undo buffer.
+    /// Set the undo buffer.
+    ///
+    /// Default is an undo-buffer with 99 undoes.
+    ///
+    /// Adjacent edits will be merged automatically into a single undo.
+    /// (Type multiple characters, they will be undone in one go.)
     #[inline]
     pub fn set_undo_buffer(&mut self, undo: Option<impl UndoBuffer + 'static>) {
         match undo {
@@ -737,19 +763,23 @@ impl TextAreaState {
         }
     }
 
-    /// Undo
+    /// Access the undo buffer.
     #[inline]
     pub fn undo_buffer(&self) -> Option<&dyn UndoBuffer> {
         self.value.undo_buffer()
     }
 
-    /// Undo
+    /// Access the undo buffer.
     #[inline]
     pub fn undo_buffer_mut(&mut self) -> Option<&mut dyn UndoBuffer> {
         self.value.undo_buffer_mut()
     }
 
     /// Begin a sequence of changes that should be undone in one go.
+    ///
+    /// Call begin_undo_seq(), then call any edit-functions. When
+    /// you are done call end_undo_seq(). Any changes will be undone
+    /// in a single step.
     #[inline]
     pub fn begin_undo_seq(&mut self) {
         self.value.begin_undo_seq()
@@ -761,25 +791,30 @@ impl TextAreaState {
         self.value.end_undo_seq()
     }
 
-    /// Get all recent replay recordings.
+    /// Get all recent replay recordings. This log can be sent
+    /// to a second TextAreaState and can be applied with replay_log().
+    ///
+    /// There are some [caveats](UndoBuffer::enable_replay_log).
     #[inline]
     pub fn recent_replay_log(&mut self) -> Vec<UndoEntry> {
         self.value.recent_replay_log()
     }
 
     /// Apply the replay recording.
+    ///
+    /// There are some [caveats](UndoBuffer::enable_replay_log).
     #[inline]
     pub fn replay_log(&mut self, replay: &[UndoEntry]) {
         self.value.replay_log(replay)
     }
 
-    /// Undo operation
+    /// Do one undo.
     #[inline]
     pub fn undo(&mut self) -> bool {
         self.value.undo()
     }
 
-    /// Redo operation
+    /// Do one redo.
     #[inline]
     pub fn redo(&mut self) -> bool {
         self.value.redo()
@@ -802,7 +837,7 @@ impl TextAreaState {
         self.value.set_styles(styles);
     }
 
-    /// Add a style for a [TextRange].
+    /// Add a style for a byte-range.
     ///
     /// The style-idx refers to one of the styles set with the widget.
     /// Missing styles are just ignored.
@@ -821,7 +856,7 @@ impl TextAreaState {
         Ok(())
     }
 
-    /// Remove the exact TextRange and style.
+    /// Remove the exact byte-range and style.
     #[inline]
     pub fn remove_style(&mut self, range: Range<usize>, style: usize) {
         self.value.remove_style(range, style);
@@ -869,9 +904,8 @@ impl TextAreaState {
 
     /// Set the offset for scrolling.
     ///
-    /// The offset uses usize for consistency, but it shouldn't
-    /// exceed u32::MAX.
-    /// TODO: details
+    /// The offset uses usize, but it shouldn't exceed (u32::MAX, u32::MAX).
+    /// This is due to the internal ScrollState that only knows usize.
     #[inline]
     pub fn set_offset(&mut self, offset: (usize, usize)) -> bool {
         self.scroll_to_cursor = false;
@@ -880,10 +914,17 @@ impl TextAreaState {
         r || c
     }
 
-    /// Offset into the first row when text-wrapping is active.
-    /// This allows displaying only part of the first text-row.
+    /// Scrolling uses the offset() for the start of the displayed text.
+    /// When text-wrapping is active, this is not enough. This gives
+    /// an extra offset into the first row where rendering should start.
     ///
-    /// This value will be trimmed before use.
+    /// You probably don't want to set this value. Use set_cursor()
+    /// instead, it will automatically scroll to make the cursor visible.
+    ///
+    /// If you really want, pos_to_line_start() can help to find the
+    /// start-position of a visual row. The x-value of the returned
+    /// TextPosition is a valid value for this function.
+    ///
     pub fn set_sub_row_offset(&mut self, sub_row_offset: upos_type) -> bool {
         self.scroll_to_cursor = false;
         let old = self.sub_row_offset;
@@ -891,8 +932,10 @@ impl TextAreaState {
         sub_row_offset != old
     }
 
-    /// Returns the offset into the first row, when text-wrapping
-    /// is active.
+    /// Returns the extra offset into the first row where rendering
+    /// starts. This is only valid if text-wrapping is active.
+    ///
+    /// Returns the index of the column.
     pub fn sub_row_offset(&self) -> upos_type {
         self.sub_row_offset
     }
@@ -901,7 +944,7 @@ impl TextAreaState {
     /// all trimmed to upos_type. sub_row_offset will only have a value if
     /// there is some text-wrapping active. hscroll.offset will only have
     /// an offset if there is *no* text-wrapping active.
-    pub fn clean_offset(&self) -> (upos_type, upos_type, upos_type) {
+    fn clean_offset(&self) -> (upos_type, upos_type, upos_type) {
         let ox = self.hscroll.offset as upos_type;
         let mut oy = self.vscroll.offset as upos_type;
 
@@ -989,13 +1032,13 @@ impl TextAreaState {
         self.value.is_empty()
     }
 
-    /// Borrow the rope
+    /// Access the underlying rope.
     #[inline]
     pub fn rope(&self) -> &Rope {
         self.value.text().rope()
     }
 
-    /// Text value
+    /// Copy of the text-value.
     #[inline]
     pub fn text(&self) -> String {
         self.value.text().string()
@@ -1049,7 +1092,7 @@ impl TextAreaState {
         self.value.line_width(row)
     }
 
-    /// Line as RopeSlice.
+    /// Line as Cow<str>.
     /// This contains the \n at the end.
     ///
     /// Panics for an invalid row.
@@ -1058,7 +1101,7 @@ impl TextAreaState {
         self.value.line_at(row).expect("valid_row")
     }
 
-    /// Line as RopeSlice.
+    /// Line as Cow<str>.
     /// This contains the \n at the end.
     #[inline]
     pub fn try_line_at(&self, row: upos_type) -> Result<Cow<'_, str>, TextError> {
@@ -1244,6 +1287,7 @@ impl TextAreaState {
     }
 
     /// Set the text value.
+    ///
     /// Resets all internal state.
     #[inline]
     pub fn set_text<S: AsRef<str>>(&mut self, s: S) {
@@ -1251,6 +1295,7 @@ impl TextAreaState {
         self.vscroll.set_offset(0);
         self.hscroll.set_offset(0);
         self.set_sub_row_offset(0);
+        self.move_col = None;
 
         self.value.set_text(TextRope::new_text(s.as_ref()));
     }
@@ -1263,6 +1308,7 @@ impl TextAreaState {
         self.vscroll.set_offset(0);
         self.hscroll.set_offset(0);
         self.set_sub_row_offset(0);
+        self.move_col = None;
 
         self.value.set_text(TextRope::new_rope(r));
     }
@@ -1270,9 +1316,13 @@ impl TextAreaState {
     /// Insert a character at the cursor position.
     /// Removes the selection and inserts the char.
     ///
-    /// This insert makes no special actions when encountering
-    /// a new-line or tab. Use insert_newline and insert_tab for
-    /// this.
+    /// You can insert a tab with this. But it will not
+    /// indent the current selection. It will expand
+    /// the tab though. Use insert_tab() for this.
+    ///
+    /// You can insert a new-line with this. But it will
+    /// not do an auto-indent.
+    /// Use insert_new_line() for this.
     pub fn insert_char(&mut self, c: char) -> bool {
         let mut insert = true;
         if self.has_selection() {